refactor(pager): Allow deque to have additional lines
to simplify the implementation.
This essentially restores the implementation at 39680f3.
So far, I tried several implementations, but this "VecDeque with cache"
version seems simplest overall.

I thought it would be simpler if `VecDeque` holds only lines that are
currently visible, but unintuitively, trying to keep only correct lines
made the logic a bit complicated.